Fixed implementation of LimitTrueDuration for state positive state with a duration less than timespan.
