+++ src/NetDaemon.Extensions.Observables.Tests/EntityExtensions.ToBooleanObservable.Scheduling.ConcreteEntity.Tests.cs
@@ -1,4 +1,3 @@
-using System.Reactive.Concurrency;
 using Moq;
 using NetDaemon.HassModel;
 using NetDaemon.HassModel.Entities;
@@ -87,7 +86,6 @@ namespace NetDaemon.Extensions.Observables.Tests
         public void WhenTrueFor_Predicate_SubscribeAfterTimeSpanPasses_True()
         {
             // Arrange
-            ChangeEntityState(Off, _lastChanged);
             var observable = _testEntity.WhenTrueFor(
                 TimeSpan.FromTicks(1),
                 s => s.IsOff(),
@@ -130,7 +128,11 @@ namespace NetDaemon.Extensions.Observables.Tests
         {
             // Arrange
             ChangeEntityState(Off, _lastChanged);
-            var observable = _testEntity.LimitTrueDuration(TimeSpan.FromMinutes(1), s => s.IsOff(), _scheduler);
+            var observable = _testEntity.LimitTrueDuration(
+                TimeSpan.FromMinutes(1),
+                s => s.IsOff(),
+                _scheduler,
+                _lastChanged.ToUniversalTime);
 
             // Act
             bool? result = null;
@@ -144,8 +146,12 @@ namespace NetDaemon.Extensions.Observables.Tests
         public void LimitTrueDuration_Predicate_LastChangedLongerAgoThanTimeSpan_False()
         {
             // Arrange
-            ChangeEntityState(Off, _lastChanged - TimeSpan.FromMinutes(2));
-            var observable = _testEntity.LimitTrueDuration(TimeSpan.FromMinutes(1), s => s.IsOff(), _scheduler);
+            ChangeEntityState(Off, _lastChanged);
+            var observable = _testEntity.LimitTrueDuration(
+                TimeSpan.FromTicks(1),
+                s => s.IsOff(),
+                _scheduler,
+                () => _lastChanged.ToUniversalTime() + TimeSpan.FromTicks(1));
 
             // Act
             bool? result = null;
@@ -159,14 +165,40 @@ namespace NetDaemon.Extensions.Observables.Tests
         public void LimitTrueDuration_Predicate_SubscribeAfterTimeSpanPasses_False()
         {
             // Arrange
-            var observable = _testEntity.LimitTrueDuration(TimeSpan.FromMinutes(1), s => s.IsOff(), _scheduler);
+            var observable = _testEntity.LimitTrueDuration(
+                TimeSpan.FromTicks(1),
+                s => s.IsOff(),
+                _scheduler,
+                _lastChanged.ToUniversalTime);
 
             // Act
             bool? result = null;
-            ChangeEntityState(Off, _lastChanged - TimeSpan.FromMinutes(2));
+            ChangeEntityState(Off, _lastChanged - TimeSpan.FromTicks(1));
             observable.Subscribe(b => result = b);
 
             // Assert
+            Assert.AreEqual(false, result);
+        }
+
+        [TestMethod]
+        public void LimitTrueDuration_Predicate_LastChangedHalfwayOfTimeSpan_FalseAfterRemainingTime()
+        {
+            ChangeEntityState(Off, _lastChanged);
+            var observable = _testEntity.LimitTrueDuration(
+                TimeSpan.FromTicks(4),
+                s => s.IsOff(),
+                _scheduler,
+                () => _lastChanged.ToUniversalTime() + TimeSpan.FromTicks(2));
+
+            bool? result = null;
+            observable.Subscribe(b => result = b);
+
+            Assert.AreEqual(true, result);
+
+            _scheduler.AdvanceBy(1);
+            Assert.AreEqual(true, result);
+
+            _scheduler.AdvanceBy(1);
             Assert.AreEqual(false, result);
         }
